Refactor: 아티스트 페이지 코드 분리 및 에러핸들링 수정
